requirements.txt
annotated-types==0.5.0
anyio==4.0.0
Babel==2.12.1
blinker==1.6.2
certifi==2023.7.22
click==8.1.7
deprecation==2.1.0
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
flask==2.3.3
Flask-BabelEx==0.9.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-WTF==1.1.1
gotrue==1.1.1
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
passlib==1.7.4
postgrest==0.10.8
pydantic==2.3.0
pydantic-core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
realtime==1.0.0
six==1.16.0
sniffio==1.3.0
speaklater==1.3
storage3==0.5.4
StrEnum==0.4.15
supabase==1.0.4
supafunc==0.2.3
typing-extensions==4.8.0
websockets==10.4
werkzeug==2.3.7
WTForms==3.0.1
zipp==3.17.0
